Reorganize code to minimize size of main
