+++ lib/doc/godoc-view.js
@@ -4,7 +4,7 @@
 import etch from 'etch'
 import EtchComponent from './../etch-component'
 
-export default class GodocPanelView extends EtchComponent {
+export default class GodocView extends EtchComponent {
   constructor (props) {
     super(props)
     if (props.model) {
