+++ lib/doc/godoc.js
@@ -6,9 +6,8 @@ import GodocPanel from './godoc-panel'
 import {isValidEditor} from '../utils'
 
 class Godoc {
-  constructor (goconfig, goget) {
+  constructor (goconfig) {
     this.goconfig = goconfig
-    this.goget = goget
     this.subscriptions = new CompositeDisposable()
     this.panelModel = new GodocPanel()
     this.subscriptions.add(this.panelModel)
@@ -54,42 +53,7 @@ class Godoc {
       if (cmd) {
         return cmd
       }
-      // first check for Go 1.6+, if we don't have that, don't even offer to
-      // 'go get', since it will definitely fail
-      return this.goconfig.locator.runtime().then((runtime) => {
-        if (!runtime) {
-          return false
-        }
-        const components = runtime.semver.split('.')
-        if (!components || components.length < 2) {
-          return false
-        }
-        const minor = parseInt(components[1], 10)
-        if (minor < 6) {
-          atom.notifications.addError('godoc requires Go 1.6 or later', {
-            detail: 'The go-plus package uses the `gogetdoc` tool, which requires Go 1.6 or later; please update your Go installation to use the doc feature.',
-            dismissable: true
-          })
-          return false
-        }
-        if (!this.goget) {
-          return false
-        }
-        return this.goget.get({
-          name: 'gogetdoc',
-          packageName: 'gogetdoc',
-          packagePath: 'github.com/zmb3/gogetdoc',
-          type: 'missing'
-        }).then((r) => {
-          if (r.success) {
-            return this.goconfig.locator.findTool('gogetdoc')
-          }
-          console.log('gogetdoc is not available and could not be installed via "go get -u github.com/zmb3/gogetdoc"; please manually install it to enable doc functionality')
-          return false
-        }).catch((e) => {
-          console.log(e)
-        })
-      })
+      return false
     })
   }
