+++ lib/format/formatter.js
@@ -5,9 +5,8 @@ import path from 'path'
 import {getEditor, isValidEditor, projectPath} from '../utils'
 
 class Formatter {
-  constructor (goconfig, goget) {
+  constructor (goconfig) {
     this.goconfig = goconfig
-    this.goget = goget
     this.subscriptions = new CompositeDisposable()
     this.saveSubscriptions = new CompositeDisposable()
     this.updatingFormatterCache = false
@@ -26,10 +25,8 @@ class Formatter {
       this.saveSubscriptions.dispose()
     }
     this.saveSubscriptions = null
-    this.goget = null
     this.goconfig = null
     this.tool = null
-    this.toolCheckComplete = null
     this.formatterCache = null
     this.updatingFormatterCache = null
     this.toolLocations = null
@@ -73,7 +70,7 @@ class Formatter {
       if (this.toolCheckComplete) {
         this.toolCheckComplete[formatTool] = false
       }
-      this.checkForTool(formatTool)
+      this.updateFormatterCache()
     }))
     this.subscriptions.add(atom.config.observe('go-plus.format.formatOnSave', (formatOnSave) => {
       if (this.saveSubscriptions) {
@@ -183,46 +180,6 @@ class Formatter {
     return false
   }
 
-  checkForTool (toolName = this.tool) {
-    if (!this.ready()) {
-      return
-    }
-    return this.goconfig.locator.findTool(toolName).then((cmd) => {
-      if (cmd) {
-        return this.updateFormatterCache().then(() => {
-          return cmd
-        })
-      }
-
-      if (!this.toolCheckComplete) {
-        this.toolCheckComplete = { }
-      }
-
-      if (!cmd && !this.toolCheckComplete[toolName]) {
-        if (!this.goget) {
-          return false
-        }
-        this.toolCheckComplete[toolName] = true
-
-        const packagePath = this.toolLocations[toolName]
-        if (packagePath) {
-          this.goget.get({
-            name: 'gofmt',
-            packageName: toolName,
-            packagePath: packagePath,
-            type: 'missing'
-          }).then(() => {
-            return this.updateFormatterCache()
-          }).catch((e) => {
-            console.log(e)
-          })
-        }
-      }
-
-      return false
-    })
-  }
-
   format (editor = getEditor(), tool = this.tool, filePath) {
     if (!isValidEditor(editor) || !editor.getBuffer()) {
       return
@@ -234,7 +191,7 @@ class Formatter {
 
     const formatCmd = this.cachedToolPath(tool, editor)
     if (!formatCmd) {
-      this.checkForTool(tool)
+      this.updateFormatterCache()
       return
     }
