+++ lib/rename/gorename.js
@@ -6,9 +6,8 @@ import RenameDialog from './rename-dialog'
 import {isValidEditor} from '../utils'
 
 class Gorename {
-  constructor (goconfig, goget) {
+  constructor (goconfig) {
     this.goconfig = goconfig
-    this.goget = goget
     this.subscriptions = new CompositeDisposable()
     this.subscriptions.add(atom.commands.add(
       'atom-text-editor', 'golang:gorename',
@@ -16,14 +15,19 @@ class Gorename {
     ))
   }
 
+  dispose () {
+    this.subscriptions.dispose()
+    this.subscriptions = null
+    this.goconfig = null
+  }
+
   commandInvoked () {
     const editor = atom.workspace.getActiveTextEditor()
     if (!isValidEditor(editor)) {
       return
     }
-    this.checkForTool(editor).then((cmd) => {
+    this.goconfig.locator.findTool('gorename').then((cmd) => {
       if (!cmd) {
-        // TODO: Show a notification?
         return
       }
 
@@ -136,51 +140,6 @@ class Gorename {
       atom.notifications.addSuccess(message.trim())
       return {success: true, result: r}
     })
-  }
-
-  checkForTool (editor) {
-    if (!this.goconfig) {
-      return Promise.resolve(false)
-    }
-
-    return this.goconfig.locator.findTool('gorename').then((cmd) => {
-      if (cmd) {
-        return cmd
-      }
-
-      if (!this.goget) {
-        return false
-      }
-
-      if (this.toolCheckComplete) {
-        return false
-      }
-
-      this.toolCheckComplete = true
-      return this.goget.get({
-        name: 'gorename',
-        packageName: 'gorename',
-        packagePath: 'golang.org/x/tools/cmd/gorename',
-        type: 'missing'
-      }).then((r) => {
-        if (r.success) {
-          return this.goconfig.locator.findTool('gorename')
-        }
-
-        console.log('gorename is not available and could not be installed via "go get -u golang.org/x/tools/cmd/gorename"; please manually install it to enable gorename behavior.')
-        return false
-      }).catch((e) => {
-        console.log(e)
-        return false
-      })
-    })
-  }
-
-  dispose () {
-    this.subscriptions.dispose()
-    this.subscriptions = null
-    this.goconfig = null
-    this.goget = null
   }
 }
