+++ lib/test/tester.js
@@ -12,6 +12,7 @@ import {getEditor, isValidEditor} from '../utils'
 
 class Tester {
   constructor (goconfig, testPanelManager) {
+    this.disposed = false
     this.goconfig = goconfig
     this.testPanelManager = testPanelManager
     this.subscriptions = new CompositeDisposable()
@@ -25,6 +26,7 @@ class Tester {
   }
 
   dispose () {
+    this.disposed = true
     this.running = true
     this.removeTempDir()
     this.clearMarkersFromEditors()
@@ -256,7 +258,7 @@ class Tester {
         }
         cover = cmd
       }).then(() => {
-        if (!go || !cover) {
+        if (!go || !cover || this.disposed) {
           this.running = false
           return
         }
@@ -269,13 +271,12 @@ class Tester {
         if (atom.config.get('go-plus.test.runTestsWithVerboseFlag')) {
           args.push('-v')
         }
-        this.testPanelManager.update({output: 'Running go ' + args.join(' ') + ' with a ' + atom.config.get('go-plus.test.timeout') + 'ms timeout', state: 'pending', exitcode: 0})
+        if (this.testPanelManager) {
+          this.testPanelManager.update({output: 'Running go ' + args.join(' ') + ' with a ' + atom.config.get('go-plus.test.timeout') + 'ms timeout', state: 'pending', exitcode: 0})
+        }
         const executorOptions = this.goconfig.executor.getOptions('file')
         executorOptions.timeout = atom.config.get('go-plus.test.timeout')
         return this.goconfig.executor.exec(cmd, args, executorOptions).then((r) => {
-          if (!this.testPanelManager) {
-            return
-          }
           let output = r.stdout
           if (r.stderr && r.stderr.trim() !== '') {
             output = r.stderr + os.EOL + r.stdout
@@ -284,9 +285,13 @@ class Tester {
           if (r.exitcode === 0) {
             this.ranges = parser.ranges(this.coverageFile)
             this.addMarkersToEditors()
-            this.testPanelManager.update({exitcode: r.exitcode, output: output.trim(), state: 'success'})
+            if (this.testPanelManager) {
+              this.testPanelManager.update({exitcode: r.exitcode, output: output.trim(), state: 'success'})
+            }
           } else {
-            this.testPanelManager.update({exitcode: r.exitcode, output: output.trim(), state: 'fail'})
+            if (this.testPanelManager) {
+              this.testPanelManager.update({exitcode: r.exitcode, output: output.trim(), state: 'fail'})
+            }
           }
 
           this.running = false
