+++ spec/doc/godoc-spec.js
@@ -14,8 +14,9 @@ describe('godoc', () => {
   let target = null
 
   beforeEach(() => {
-    lifecycle.setup()
     runs(() => {
+      lifecycle.setup()
+      atom.packages.triggerDeferredActivationHooks()
       gopath = fs.realpathSync(lifecycle.temp.mkdirSync('gopath-'))
       process.env.GOPATH = gopath
     })
@@ -27,9 +28,11 @@ describe('godoc', () => {
     runs(() => {
       let pack = atom.packages.loadPackage('go-plus')
       pack.activateNow()
+      atom.packages.triggerActivationHook('core:loaded-shell-environment')
+      atom.packages.triggerActivationHook('language-go:grammar-used')
       mainModule = pack.mainModule
-      mainModule.getGoconfig()
-      mainModule.getGoget()
+      mainModule.provideGoConfig()
+      mainModule.provideGoGet()
       mainModule.loadDoc()
     })
